feat: implement product's breadcrumbs
