+++ client-app/core/api/graphql/catalog/queries/getProduct/getProductQuery.graphql
@@ -10,6 +10,8 @@ query GetProduct($storeId: String!, $currencyCode: String!, $cultureName: String
       url
     }
     breadcrumbs {
+      itemId
+      typeName
       title
       seoPath
     }
