+++ client-app/core/api/graphql/types.ts
@@ -3250,7 +3250,7 @@ export type GetProductQueryVariables = Exact<{
 }>;
 
 
-export type GetProductQuery = { product?: { name: string, id: string, code: string, slug?: string | null | undefined, outline?: string | null | undefined, imgSrc?: string | null | undefined, images?: Array<{ url?: string | null | undefined } | null | undefined> | null | undefined, breadcrumbs?: Array<{ title: string, seoPath?: string | null | undefined } | null | undefined> | null | undefined, description?: { content: string, id: string, languageCode: string, reviewType: string } | null | undefined, descriptions?: Array<{ content: string, id: string, languageCode: string, reviewType: string } | null | undefined> | null | undefined, properties?: Array<{ name: string, value?: string | null | undefined, type?: string | null | undefined } | null | undefined> | null | undefined, variations?: Array<{ id?: string | null | undefined, name?: string | null | undefined, code?: string | null | undefined, images?: Array<{ url?: string | null | undefined } | null | undefined> | null | undefined, properties?: Array<{ name: string, value?: string | null | undefined, type?: string | null | undefined } | null | undefined> | null | undefined, availabilityData?: { isActive?: boolean | null | undefined, isAvailable?: boolean | null | undefined, isBuyable?: boolean | null | undefined, isInStock?: boolean | null | undefined, availableQuantity: any } | null | undefined, price?: { actual?: { amount: any, formattedAmount: string } | null | undefined, discountAmount?: { amount: any, formattedAmount: string } | null | undefined, sale?: { amount: any, formattedAmount: string } | null | undefined, list?: { amount: any, formattedAmount: string } | null | undefined } | null | undefined } | null | undefined> | null | undefined, availabilityData?: { isActive?: boolean | null | undefined, isAvailable?: boolean | null | undefined, isBuyable?: boolean | null | undefined, isInStock?: boolean | null | undefined, availableQuantity: any } | null | undefined, price?: { actual?: { amount: any, formattedAmount: string } | null | undefined, discountAmount?: { amount: any, formattedAmount: string } | null | undefined, sale?: { amount: any, formattedAmount: string } | null | undefined, list?: { amount: any, formattedAmount: string } | null | undefined } | null | undefined } | null | undefined };
+export type GetProductQuery = { product?: { name: string, id: string, code: string, slug?: string | null | undefined, outline?: string | null | undefined, imgSrc?: string | null | undefined, images?: Array<{ url?: string | null | undefined } | null | undefined> | null | undefined, breadcrumbs?: Array<{ itemId: string, typeName?: string | null | undefined, title: string, seoPath?: string | null | undefined } | null | undefined> | null | undefined, description?: { content: string, id: string, languageCode: string, reviewType: string } | null | undefined, descriptions?: Array<{ content: string, id: string, languageCode: string, reviewType: string } | null | undefined> | null | undefined, properties?: Array<{ name: string, value?: string | null | undefined, type?: string | null | undefined } | null | undefined> | null | undefined, variations?: Array<{ id?: string | null | undefined, name?: string | null | undefined, code?: string | null | undefined, images?: Array<{ url?: string | null | undefined } | null | undefined> | null | undefined, properties?: Array<{ name: string, value?: string | null | undefined, type?: string | null | undefined } | null | undefined> | null | undefined, availabilityData?: { isActive?: boolean | null | undefined, isAvailable?: boolean | null | undefined, isBuyable?: boolean | null | undefined, isInStock?: boolean | null | undefined, availableQuantity: any } | null | undefined, price?: { actual?: { amount: any, formattedAmount: string } | null | undefined, discountAmount?: { amount: any, formattedAmount: string } | null | undefined, sale?: { amount: any, formattedAmount: string } | null | undefined, list?: { amount: any, formattedAmount: string } | null | undefined } | null | undefined } | null | undefined> | null | undefined, availabilityData?: { isActive?: boolean | null | undefined, isAvailable?: boolean | null | undefined, isBuyable?: boolean | null | undefined, isInStock?: boolean | null | undefined, availableQuantity: any } | null | undefined, price?: { actual?: { amount: any, formattedAmount: string } | null | undefined, discountAmount?: { amount: any, formattedAmount: string } | null | undefined, sale?: { amount: any, formattedAmount: string } | null | undefined, list?: { amount: any, formattedAmount: string } | null | undefined } | null | undefined } | null | undefined };
 
 export type CategoriesQueryVariables = Exact<{
   storeId: Scalars['String'];
