+++ client-app/shared/catalog/composables/useProducts.ts
@@ -1,8 +1,10 @@
-import { Ref, ref, computed } from "vue";
+import { Ref, ref, computed, watch } from "vue";
 import { searchProducts, searchRelatedProducts, getProduct } from "@core/api/graphql/catalog";
 import { Product } from "@core/api/graphql/types";
 import { Logger } from "@core/utilities";
 import { ProductsSearchParams } from "../types";
+import { useCart } from "@/shared/cart";
+import _ from "lodash";
 
 export default () => {
   const products: Ref<Product[]> = ref([]);
@@ -11,6 +13,10 @@ export default () => {
   const total: Ref<number> = ref(0);
   const pages: Ref<number> = ref(0);
   const loading: Ref<boolean> = ref(true);
+  const withVariations = computed(() => product.value.variations?.length);
+  const variationsCartTotal = ref(0);
+
+  const { loading: cartLoading, getItemsTotal } = useCart();
 
   async function fetchRelatedProducts(id: string) {
     loading.value = true;
@@ -54,6 +60,22 @@ export default () => {
     }
   }
 
+  watch(
+    () => cartLoading.value === false && loading.value === false,
+    (condition) => {
+      if (condition && product.value.variations?.length) {
+        const variationsIds = _(product.value.variations)
+          .filter((x) => !!x?.id)
+          .map((x) => x?.id as string)
+          .value();
+
+        variationsIds.push(product.value.id);
+
+        variationsCartTotal.value = getItemsTotal(variationsIds);
+      }
+    }
+  );
+
   return {
     fetchProducts,
     loadProduct,
@@ -64,5 +86,7 @@ export default () => {
     total: computed(() => total.value),
     pages: computed(() => pages.value),
     loading: computed(() => loading.value),
+    withVariations: computed(() => withVariations.value),
+    variationsCartTotal: computed(() => variationsCartTotal.value),
   };
 };
